Fix tree bugs as per retesting doc
