+++ lib/SoftwareOffline.py
@@ -10,7 +10,8 @@ from .SoftwareUtil import *
 from .SoftwareFileDataManager import *
 from .SoftwareModels import SessionSummary, KeystrokeAggregate
 from .SoftwareWallClock import *
-from .SoftwareDashboard import *
+from .SoftwareFileChangeInfoSummaryData import *
+from .SoftwarePayload import *
 
 # Constants
 SERVICE_NOT_AVAIL = "Our service is temporarily unavailable.\n\nPlease try again later.\n"
@@ -74,7 +75,7 @@ def getSessionThresholdSeconds():
     thresholdSeconds = getItem('sessionThresholdInSec') or DEFAULT_SESSION_THRESHOLD_SECONDS
     return thresholdSeconds
 
-def clearSessionSummaryDataData():
+def clearSessionSummaryData():
     log('--- clearing session summary data ---')
     emptyData = SessionSummary()
     saveSessionSummaryToDisk(emptyData)
@@ -96,89 +97,8 @@ def getMinutesSinceLastPayload():
 
         if diffInSec > 0 and diffInSec < getSessionThresholdSeconds():
             minutesSinceLastPayload = diffInSec / 60
-    else:
-        refreshSessionSummaryTimer = Timer(1.0, getSessionSummaryStatus)
-        refreshSessionSummaryTimer.start()
 
     return minutesSinceLastPayload
-
-# TODO: this method hangs for a while (bc no cache now?)
-def launchCodeTimeMetrics():
-    fetchCodeTimeMetricsDashboard()
-    file = getDashboardFile()
-    sublime.active_window().open_file(file)
-
-
-def fetchCodeTimeMetricsDashboard():
-    serverOnline = serverIsAvailable()
-    summaryInfoFile = getSummaryInfoFile()
-
-    if serverOnline:
-        # fetch the backend data
-        islinux = "true"
-        if isWindows() is True or isMac() is True:
-            islinux = "false"
-
-        # TODO: find sublime setting for showGitMEtrics and replace true with it
-        showGitMetrics = True 
-        api = '/dashboard?showGit=' + 'true' + '&linux=' + islinux + '&showToday=false'
-        response = requestIt("GET", api, None, getItem("jwt"))
-
-        summaryContent = ""
-        try:
-            summaryContent = response.read().decode('utf-8')
-        except Exception as ex:
-            summaryContent = SERVICE_NOT_AVAIL
-            log("Code Time: Unable to read response data: %s" % ex)
-
-        try:
-            with open(summaryInfoFile, 'w', encoding='utf-8') as f:
-                f.write(summaryContent)
-        except Exception as ex:
-            log("Code Time: Unable to write dashboard summary content: %s" % ex)
-
-    # concat summary info with the dashboard file
-    dashboardFile = getDashboardFile()
-    dashboardContent = ""
-
-    d = datetime.datetime.now()
-
-    formattedDate = d.strftime("%a %b %d %I:%M %p")
-    dashboardContent += "CODE TIME          (Last updated on %s)\n\n" % formattedDate
-
-    formattedTodayDate = d.strftime("%a %b %d")
-    todayHeader = "Today (%s)" % formattedTodayDate
-    dashboardContent += getSectionHeader(todayHeader)
-
-    summary = getSessionSummaryStatus()
-    if (summary is not None):
-        averageTime = getAverageDailyTime(summary)["formatted"]
-        hoursCodedToday = getCurrentDayTime(summary)["formatted"]
-
-        liveshareTime = None 
-        if summary['liveshareMinutes']:
-            liveshareTime = humanizeMinutes(summary['liveshareMinutes'])
-
-        currentEditorMinutesStr = getHumanizedWcTime()
-        dashboardContent += getDashboardRow('Editor time today', currentEditorMinutesStr)
-        dashboardContent += getDashboardRow("Code time today", hoursCodedToday)
-        dashboardContent += getDashboardRow("90-day avg", averageTime)
-        if liveshareTime:
-            dashboardContent += getDashboardRow('Live Share', liveshareTime)
-        dashboardContent += "\n"
-
-    if (os.path.exists(summaryInfoFile)):
-        try:
-            with open(summaryInfoFile, 'r', encoding="utf-8") as summaryInfoFileContent:
-                dashboardContent += summaryInfoFileContent.read()
-        except Exception as ex:
-            log("Code Time: Unable to read summary info file content: %s" % ex)
-
-    try:
-        with open(dashboardFile, 'w', encoding='utf-8') as f:
-            f.write(dashboardContent)
-    except Exception as ex:
-        log("Code Time: Unable to write local dashboard content: %s" % ex)
 
 # store the payload offline...
 def storePayload(payload):
@@ -253,57 +173,6 @@ def storePayload(payload):
             dsFile.write(json.dumps(payload) + "\n")
     except Exception as ex:
         log('Error appending to the Software data store file: %s' % ex)
-
-# send the data that has been saved offline
-def sendOfflineData():
-    batchSendData('/data/batch', getSoftwareDataStoreFile())
-
-def sendOfflineEvents():
-    batchSendData('/data/event', getPluginEventsFile())
-
-def sendOfflineTimeData():
-    batchSendData('/data/time', getTimeDataSummaryFile(), isArray=True)
-
-
-def batchSendData(api, file, isArray=False):
-    isOnline = serverIsAvailable()
-    if not isOnline:
-        return 
-
-    try:
-        # print('batch sending {}'.format(file))
-        if os.path.exists(file):
-            payloads = None 
-            if isArray:
-                payloads = getFileDataArray(file)
-            else:
-                payloads = getFileDataPayloadsAsJson(file)
-            batchSendPayloadData(api, file, payloads)
-    except Exception as ex:
-        log('Error batch sending payloads: %s' % ex)
-
-            
-
-def batchSendPayloadData(api, file, payloads):
-    if (payloads is not None and len(payloads) > 0):
-        log('sending batch payloads')
-
-        # go through the payloads array 50 at a time
-        batch = []
-        length = len(payloads)
-        for i in range(length):
-            payload = payloads[i]
-            if (len(batch) >= 50):
-                requestIt("POST", "/data/batch", json.dumps(batch), getItem("jwt"))
-                # send batch
-                batch = []
-            batch.append(payload)
-
-        # send remaining batch
-        if (len(batch) > 0):
-            requestIt("POST", "/data/batch", json.dumps(batch), getItem("jwt"))
-        
-        os.remove(file)
